requirements/common.txt
Django==1.5.4
South==0.8.2
requests==1.2.3
django-mathfilters==0.1.3
osmread==0.1
feedparser==5.1.3
python-dateutil==2.1
psycopg2==2.5.1
django-registration==1.0
boto==2.13.3
git+git://github.com/jpulec/PyQS.git
git+git://github.com/jpulec/django-ratings.git
django-organizations==0.1.8
